+++ docs/modules/events.md
@@ -2,39 +2,43 @@
 
 ## Назначение
 
-Events отвечает за отложенную функциональность мероприятий Procollab: список
-мероприятий, карточку мероприятия, тип проведения, теги, зарегистрированных
-пользователей и опциональную служебную автопубликацию в Telegram.
+Events отвечает за отложенную функциональность мероприятий Procollab. Модуль
+сохраняет модели, admin-управление, связанный список мероприятий пользователя и
+опциональную служебную автопубликацию в Telegram, но публичные endpoints
+`/events/` сейчас не подключены в корневом URLConf.
 
 ## Статус модуля
 
 Модуль разрабатывался под задачу, которая перестала быть актуальной в процессе
-работы. Сейчас он не является активным продуктовым направлением, но его базовый
-read/admin-контур поддерживается и покрыт regression-тестами.
+работы. Сейчас он не является активным продуктовым направлением. Чтобы
+неиспользуемые ручки не торчали наружу и не попадали в Swagger, `events.urls`
+не подключается в `procollab/urls.py`.
 
 Часть возможностей заложена в модели, но не имеет публичных endpoints. Это
 технический долг, который нужно решать только при возвращении мероприятий в
 активную продуктовую работу.
 
-Текущие тесты покрывают основные API-flow, модель лайков, registered users,
-Telegram helpers и autoposting signal. Coverage по модулю без tests/migrations
-составляет около 98%.
+Текущие тесты покрывают отключение публичных `/events/` endpoints, связанный
+endpoint пользователя, модель лайков, Telegram helpers и autoposting signal.
 
 ## Основные возможности
 
-- просмотр списка мероприятий;
-- просмотр карточки мероприятия;
-- создание, обновление и удаление мероприятия staff-пользователем;
-- фильтрация мероприятий по title, text и дате создания;
+- хранение мероприятий в базе;
+- управление мероприятиями через Django admin;
 - хранение тегов мероприятия через `django-taggit`;
 - хранение зарегистрированных пользователей мероприятия;
-- получение списка зарегистрированных пользователей;
-- получение списка типов мероприятий;
+- получение списка мероприятий текущего пользователя через users endpoint;
 - опциональная автопубликация мероприятия в Telegram при настроенных
   `AUTOPOSTING_ON`, `TELEGRAM_BOT_TOKEN` и `TELEGRAM_CHANNEL`.
 
 Заложено в модели, но сейчас не имеет отдельного публичного API:
 
+- список мероприятий;
+- карточка мероприятия;
+- создание, обновление и удаление мероприятия через `/events/`;
+- фильтрация мероприятий;
+- список зарегистрированных пользователей мероприятия;
+- список типов мероприятий;
 - избранное через `favorites`;
 - просмотры через `views`;
 - лайки через `LikesOnEvent`;
@@ -43,8 +47,8 @@ Telegram helpers и autoposting signal. Coverage по модулю без tests/
 ## Архитектура
 
 - `events/models.py` - модели `Event` и `LikesOnEvent`.
-- `events/views.py` - HTTP endpoints для списка, деталей, зарегистрированных
-  пользователей и типов мероприятий.
+- `events/views.py` - неактивные HTTP endpoints для списка, деталей,
+  зарегистрированных пользователей и типов мероприятий.
 - `events/serializers.py` - request/response serializers для списка, деталей и
   зарегистрированных пользователей.
 - `events/filters.py` - фильтры списка мероприятий.
@@ -52,7 +56,8 @@ Telegram helpers и autoposting signal. Coverage по модулю без tests/
 - `events/helpers.py` - низкоуровневые вызовы Telegram API.
 - `events/signals.py` - autoposting signal для Telegram.
 - `events/admin.py` - настройка Django admin.
-- `events/tests/` - regression-тесты модуля.
+- `events/tests/` - regression-тесты отключенных публичных URL, связанного users
+  endpoint, модели лайков и Telegram-интеграции.
 
 ## Основные сущности
 
@@ -65,36 +70,25 @@ Telegram helpers и autoposting signal. Coverage по модулю без tests/
 
 ## API
 
-- `GET /events/` - список мероприятий.
-- `POST /events/` - создание мероприятия staff-пользователем.
-- `GET /events/<id>/` - детали мероприятия.
-- `PUT /events/<id>/` - полное обновление мероприятия staff-пользователем.
-- `PATCH /events/<id>/` - частичное обновление мероприятия staff-пользователем.
-- `DELETE /events/<id>/` - удаление мероприятия staff-пользователем.
-- `GET /events/<id>/registered/` - список зарегистрированных пользователей.
-- `GET /events/types/` - список типов мероприятий.
+Публичные endpoints модуля Events сейчас отключены на уровне
+`procollab/urls.py`. Эти URLs не должны быть доступны извне:
 
-Связанный endpoint:
+- `/events/`;
+- `/events/<id>/`;
+- `/events/<id>/registered/`;
+- `/events/types/`.
+
+Активный связанный endpoint:
 
 - `GET /auth/users/current/events/` - мероприятия, на которые зарегистрирован
   текущий пользователь.
 
 ## Основные сценарии
 
-### 1. Просмотр мероприятий
+### 1. Управление мероприятиями
 
-Пользователь открывает список или карточку мероприятия. Read endpoints доступны
-без staff-прав, потому что используется permission `IsStaffOrReadOnly`.
-
-Список мероприятий можно фильтровать по:
-
-- `title__contains`;
-- `text__contains`;
-- `datetime_created__gt`.
-
-### 2. Управление мероприятиями
-
-Создание, обновление и удаление мероприятия доступны только staff-пользователям.
+Пока публичный API отключен, создание, обновление и удаление мероприятий
+остаются admin-сценарием.
 
 При создании и обновлении используются поля:
 
@@ -109,17 +103,14 @@ Telegram helpers и autoposting signal. Coverage по модулю без tests/
 - теги;
 - избранное.
 
-### 3. Зарегистрированные пользователи
+### 2. Зарегистрированные пользователи
 
 Связь пользователей с мероприятием хранится в `Event.registered_users`.
-
-Endpoint `/events/<id>/registered/` возвращает список зарегистрированных
-пользователей через `RegisteredUserListSerializer`.
 
 Для текущего пользователя список его мероприятий доступен через
 `/auth/users/current/events/`.
 
-### 4. Лайки, избранное и регистрация
+### 3. Лайки, избранное и регистрация
 
 Модель поддерживает:
 
@@ -129,7 +120,7 @@ Endpoint `/events/<id>/registered/` возвращает список зарег
 Отдельных публичных endpoints для toggle-like, favorite или регистрации на
 мероприятие в `events/urls.py` сейчас нет.
 
-### 5. Telegram autoposting
+### 4. Telegram autoposting
 
 Если `AUTOPOSTING_ON = True`, при сохранении `Event` импортируются signals и
 срабатывает автопубликация:
@@ -142,8 +133,7 @@ Endpoint `/events/<id>/registered/` возвращает список зарег
 
 ## Ограничения и правила
 
-- Read endpoints доступны всем пользователям.
-- Write endpoints доступны только staff-пользователям.
+- `/events/` endpoints не подключены в корневом URLConf.
 - `text`, `cover_url` и `datetime_of_event` обязательны при создании
   мероприятия.
 - `short_text` ограничен 256 символами.
@@ -159,13 +149,7 @@ Endpoint `/events/<id>/registered/` возвращает список зарег
 
 Они проверяют:
 
-- публичный список и detail мероприятия;
-- создание, частичное обновление и удаление staff-пользователем;
-- запрет write-действий обычному и анонимному пользователю;
-- validation errors при пустом `text` и пустом payload;
-- фильтры списка мероприятий;
-- `/events/types/`;
-- `/events/<id>/registered/`;
+- недоступность публичных `/events/` URLs;
 - `LikesOnEventManager`;
 - `/auth/users/current/events/`;
 - Telegram helpers и autoposting signal.
