+++ events/urls.py
@@ -4,6 +4,8 @@ from events.views import EventsList, EventDetail, EventRegisteredUsersList, Even
 
 app_name = "events"
 
+# Файл сохранен для возможного возврата модуля мероприятий, но сейчас не
+# подключен в `procollab/urls.py`, поэтому /events/ endpoints недоступны извне.
 urlpatterns = [
     path("", EventsList.as_view()),
     path("<int:pk>/", EventDetail.as_view()),
